file exists

+++ dirtmdb.py
@@ -10,6 +10,7 @@ def uselessFile(entryName):
 
 
 def rename_dirs_with_tmdb(root_dir):
+    fo = open("sus.txt", "w")
     for index, dirname in enumerate(os.listdir(root_dir)):
         if uselessFile(dirname):
             continue
@@ -45,17 +46,26 @@ def rename_dirs_with_tmdb(root_dir):
                             new_dirname = f"{original_dirname} {{tmdb-{tmdb_id}}}"
                             new_dirpath = os.path.join( os.path.dirname(dirpath), new_dirname)
                             print(f"{index} : {original_dirname} ==> {new_dirname}")
-                            os.rename(dirpath, new_dirpath)
+                            # os.rename(dirpath, new_dirpath)
                         else:
                             # print(f"{index} : {original_dirname} origin TMDb name.")
                             new_dirname =  original_dirname.replace(m.group(0), f"tmdb-{tmdb_id}")
                             new_dirpath = os.path.join( os.path.dirname(dirpath), new_dirname)
                             print(f"{index} : {original_dirname} ==> {new_dirname}")
+                        if os.path.exists(new_dirpath):
+                            print(f"{new_dirpath} exists, skip")
+                            break
+                        try:
                             os.rename(dirpath, new_dirpath)
+                        except:
+                            print("rename exception")
+                        fo.write( new_dirname + os.linesep )
+
                         break
             if not nfo_found:
                 thedirname = os.path.basename(dirpath)
                 print(f"{index} : {thedirname} .nfo file not found")
+    fo.close()
 
 
 def rename_movie_dirs_with_tmdb(root_dir):
